Finished version, submitted

+++ ft_strlcat.c
@@ -6,7 +6,7 @@
 /*   By: jkaplin <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/02/12 18:46:07 by jkaplin           #+#    #+#             */
-/*   Updated: 2019/03/03 16:02:44 by jkaplin          ###   ########.fr       */
+/*   Updated: 2019/03/04 17:02:48 by jkaplin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,20 +14,20 @@
 
 size_t		ft_strlcat(char *dst, const char *src, size_t dstsize)
 {
-	size_t	i;
+	int		i;
 	int		j;
 	int		real_dstsize;
 
 	i = ft_strlen(dst);
 	real_dstsize = i;
 	j = 0;
-	while (src[j] && i < dstsize - 1)
+	while (src[j] && i < (int)dstsize - 1)
 	{
 		dst[i] = src[j];
 		i++;
 		j++;
 	}
-	if (i < dstsize)
+	if (i < (int)dstsize)
 		dst[i] = '\0';
 	return (ft_min(real_dstsize, dstsize) + ft_strlen(src));
 }
